An updated version of the index.html file is provided below.
When a user clicks on an article card, a modal dialog will now appear. This dialog presents three options:

Visit Substack: Opens the article link in a new browser tab.

Share: Uses the browser's native Web Share API to share the article's title and link. If the browser does not support this feature, an alert will appear with the link to be copied manually.

Cancel: Closes the dialog prompt.

The modal can also be closed by clicking the background overlay or by pressing the Escape key.